requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
motor==3.1.1
pymongo==4.4.0
pydantic>=1.10.2,<2.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv==1.0.0
email-validator==2.0.0
requests==2.31.0
